# backend/docs/openapi_users.yaml
openapi: 3.0.3
info:
  title: Spartan Market - User Profiles API
  description: API para gestión de perfiles de usuario, avatares, privacidad e historial de compras
  version: 1.0.0
  contact:
    name: Spartan Market Team
    email: support@spartanmarket.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.spartanmarket.com
    description: Production server
security:
  - FirebaseAuth: []
paths:
...
  /api/v1/users/credits/me:
    get:
      tags:
        - Credits
      summary: Obtener créditos del usuario
      description: Obtener cantidad de créditos disponibles del usuario
      security:
        - FirebaseAuth: []
      responses:
        '200':
          description: Créditos obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditsResponse'
        '401':
          description: No autorizado
  /api/v1/users/credits/buy:
    post:
      tags:
        - Credits
      summary: Comprar créditos
      description: Iniciar proceso de compra de créditos con MercadoPago
      security:
        - FirebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyCreditsRequest'
      responses:
        '200':
          description: Compra iniciada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MercadoPagoPreferenceResponse'
        '400':
          description: Datos inválidos
        '401':
          description: No autorizado

  # ========================================
  # HISTORIAL DE COMPRAS
  # ========================================
...
components:
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de autenticación de Firebase
  schemas:
...
    BuyCreditsRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: integer
          minimum: 1
          maximum: 10000
          description: Cantidad de créditos a comprar
...
    CreditsResponse:
      type: object
      properties:
        credits:
          type: integer
          description: Cantidad de créditos disponibles
        total_earned:
          type: integer
          description: Total de créditos ganados
        total_spent:
          type: integer
          description: Total de créditos gastados
    MercadoPagoPreferenceResponse:
      type: object
      properties:
        preference_id:
          type: string
          description: ID de la preferencia de MercadoPago
        init_point:
          type: string
          description: URL de pago de MercadoPago
        sandbox_init_point:
          type: string
          description: URL de pago en modo sandbox
...